refactor: modal logic 훅으로 리팩토링 (#ATR-513)
