+++ apps/service/src/features/user-setting/lib/hook/index.ts
@@ -1,1 +1,3 @@
+export { default as useCheckDuplicateName } from './useCheckDuplicateName'
+export { default as useUserInterestTag } from './useUserInterestTag'
 export { default as useUserSetting } from './useUserSetting'
