+++ apps/service/src/features/user-setting/ui/modal/UserSettingNicknameModal.tsx
@@ -1,9 +1,5 @@
-import { useDebounce } from '@/shared/lib'
 import { WarnTxt } from '@/shared/ui'
-import { useMutation } from '@tanstack/react-query'
-import { useEffect, useState } from 'react'
-import { postDuplicateName } from '../../api'
-import { checkUserSettingInputError } from '../../lib'
+import { useCheckDuplicateName } from '../../lib'
 
 interface UserInfoNicknameInputType {
   initialValue: string
@@ -14,51 +10,8 @@ export default function UserInfoNicknameInput({
   initialValue,
   setModalValue,
 }: UserInfoNicknameInputType) {
-  const [alert, setAlert] = useState('')
-  const [valid, setIsValid] = useState(false)
-  const [nickname, setNickname] = useState(initialValue)
-  const debounceDuplicateInputValue = useDebounce(nickname, 200)
-
-  const { mutate } = useMutation({
-    mutationFn: postDuplicateName,
-    onError: (err) => {
-      setAlert(err.message)
-    },
-    onSuccess: () => {
-      setAlert('')
-      setModalValue({ nickname: debounceDuplicateInputValue })
-      setIsValid(true)
-    },
-  })
-
-  useEffect(() => {
-    if (
-      debounceDuplicateInputValue !== initialValue &&
-      checkUserSettingInputError(debounceDuplicateInputValue, setAlert)
-    ) {
-      mutate({ nickname: debounceDuplicateInputValue })
-    }
-  }, [debounceDuplicateInputValue, initialValue, mutate])
-
-  useEffect(() => {
-    if (nickname.length === 0) {
-      setModalValue(undefined)
-    }
-
-    if (alert.length > 0) {
-      setModalValue(undefined)
-    }
-
-    if (initialValue === debounceDuplicateInputValue) {
-      setModalValue(undefined)
-    }
-  }, [
-    alert.length,
-    debounceDuplicateInputValue,
-    initialValue,
-    nickname.length,
-    setModalValue,
-  ])
+  const { nickname, nicknameChangeHandler, alert, valid } =
+    useCheckDuplicateName(initialValue, setModalValue)
 
   return (
     <div className="mb-6 block">
@@ -72,8 +25,7 @@ export default function UserInfoNicknameInput({
         className="block h-12 w-full rounded-lg border border-gray-100 px-3 py-2 text-base outline-none transition-colors focus:border-blue-400 dark:border-gray-700 dark:bg-gray-700 dark:focus:bg-gray-800"
         placeholder="서비스에서 사용할 닉네임을 입력해 주세요"
         onChange={(e) => {
-          setNickname(e.target.value)
-          setIsValid(false)
+          nicknameChangeHandler(e.target.value)
         }}
       />
       {!!alert.length && <WarnTxt content={alert} color="red" />}
